rename: role, status -> entity 폴더 이동
