+++ src/main/java/com/example/dp/domain/user/entity/UserRole.java
@@ -1,4 +1,4 @@
-package com.example.dp.domain.user;
+package com.example.dp.domain.user.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
