+++ src/main/java/com/example/dp/domain/user/entity/UserStatus.java
@@ -1,4 +1,4 @@
-package com.example.dp.domain.user;
+package com.example.dp.domain.user.entity;
 
 import lombok.Getter;
